consolidate walg functions to a single file
